receitasapp/requirements.txt
appdirs==1.4.0
Django==1.10.5
djangorestframework==3.5.3
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.1.10
six==1.10.0
Pillow=4.0.0
drf-nested-routers=0.11.1
coreapi=2.2.4